requirements.txt
# Core dependencies (stdlib-only except pyyaml)
pyyaml>=6.0    # optional: config file support (accounts.yaml)
